Don't build node_exporter from source on macOS
At some point, a binary distribution of node_exporter became available
for macOS, which is extremely convenient given that we are just
starting to see weird compiler errors when trying to build it from
source. At any rate, we shouldn't bother building it from source when
there's a perfectly good binary that we can use which is already
available when `go get` is run.