Remove broken build-source.yml task
We haven't used this task in ages, and we no longer provide the golang
package on our build nodes so it's effectively broken. It may still
work with a golang package present, but we have no way of verifying
this.

Anyways, the safe thing to do is to kill it. Instead, in the near
future, we'll add tasks to download pre-built node_exporter binary
packages.