Support installation from binary package
This is useful not only for systems where an Ansible package manager
module is not available, or also for systems where the latest version
is extremely old (such as debian).

At the moment, this doesn't work on Windows, but given that the author
of wmi_exporter publishes binary releases, this could be added without
too much trouble in the future.