Don't rely on launchctl list command
The `launchctl list` command doesn't always return a PID field in its
output, even if the process is running. This change uses the
conventional Unix process commands to check if the process is running
instead.

Note that this change also modifies the task flow for the
jenkins-swarm-client role. Previously we just called `launchctl load`
with `failed_when: false`, and looked at the command output to
determine if the service was already running. Now we inspect the PID
first and selectively start the service, which should also make it
safer for user to idempotently and safely call `startup.yml`.