Use "is succeeded" instead of "is success"
This matches "is failed" and "is changed", which we already use widely
in cidre. It's also the recommended terminology in Ansible:

"Older versions of Ansible used success and fail, but succeeded and
failed use the correct tense. All of these options are now valid."

From https://docs.ansible.com/ansible/latest/user_guide/playbooks_conditionals.html#conditions-based-on-registered-variables