Blow away entire file instead of using lineinfile
The problem with using lineinfile is that the config file for
prometheus-node-exporter (currently) contains a multi-line definition
for ARGS, and the regex being used by lineinfile would only replace
the first line. This would cause syntax errors when the service tried
to parse that file.

Instead, we should do the safe thing and just overwrite the entire
file.