Don't install recommended packages
On bionic (and below), there aren't any recommended packages, so this
isn't a problem. However, on focal (and above), the recommended packages
are dbus and prometheus-node-exporter-collectors. The latter isn't
necessary and should be installed by the user if desired. However, the
former causes problems in Docker images especially on focal.