Configure user/group on demand for Unix platforms
Because the user/group are used by the service files, we can't create
them only when installing the binary. However, it's not appropriate to
*always* create the user/group, because if these are defined to
`ansible_user` then the task will fail when attempting to modify a
logged-in user. So instead, this commit introduces a new variable,
node_exporter_create_user, which can be used to control when the
user/group are created.

Since the user/group are needed in most cases, the default value of this
var is true. This breaking change in behavior will be followed by a
minor version bump in a subsequent commit.