Restart the service when arguments have changed
Notify the restart handler if command line arguments have changed.

The SysVinit restart handler is now composed of two tasks, the first one
ensures that the service is running, and the second one is restarting
the service. This is necessary, because restarting the service will fail
in case the service hasn't been run at least once, due to the pid file
directory being absent.